apps/evals-leaderboard/requirements.txt
gradio
requests
pyyaml
modelscope
